+++ byconServices/api.py
@@ -2,9 +2,13 @@
 
 import json
 from humps import camelize
+from os import path
 
 from collections import OrderedDict
-from bycon import BYC, prdbug, prjsonhead, prjsontrue, select_this_server
+
+from bycon import BYC, BYC_PARS, ByconFilteringTerms, ChroNames, load_yaml_empty_fallback, prdbug, prjsonhead, prjsontrue, select_this_server
+
+services_conf_path = path.join( path.dirname( path.abspath(__file__) ), "config" )
 
 ################################################################################
 
@@ -35,13 +39,18 @@ def api():
 class ByconOpenAPI:
     def __init__(self):
         self.entity_defaults = BYC.get("entity_defaults", {})
+        self.examples = load_yaml_empty_fallback(path.join(services_conf_path, "examples.yaml"))
         self.argument_definitions = BYC.get("argument_definitions", {}).get("$defs", {})
+        self.mode = BYC_PARS.get("mode", "__none__")
         self.this_server = select_this_server()
 
         self.beacon_eps = ["info", "dataset", "cohort", "genomicVariant", "analysis", "biosample", "individual", "filteringTerm"]
         self.service_eps = ["collations", "intervalFrequencies", "geolocations", "publications"]
-
+        self.collation_types = ["NCIT", "PMID", "NCITsex", "icdom"]
         self.general_pars = ["skip", "limit", "requested_granularity"]
+
+        self.include_id_paths = True
+        self.get_parameter_values = True
 
         info = self.entity_defaults.get("info", {}).get("content", {})
 
@@ -60,6 +69,8 @@ class ByconOpenAPI:
             }
         })
 
+        self.__add_example_links()
+        self.__modify_by_mode()
         self.__add_beacon_paths()
         self.__add_service_paths()
         self.__add_component_parameters()
@@ -83,6 +94,39 @@ class ByconOpenAPI:
 
     # -------------------------------------------------------------------------#
     # ----------------------------- private -----------------------------------#
+    # -------------------------------------------------------------------------#
+
+    def __add_example_links(self):
+        e_l = ['<a href="?mode=">[Show all paths]</a>']
+        for e in self.examples.keys():
+            e_l.append(f'<a href="?mode={e}">{e} example]</a>')
+        self.oapi["info"]["description"] += f'<hr/><b>{" | ".join(e_l)}</b><hr/>'
+
+
+    # -------------------------------------------------------------------------#
+
+    def __modify_by_mode(self):
+        if not (e := self.examples.get(self.mode)):
+            self.mode = False
+            return
+
+        self.beacon_eps = e.get("entities", self.beacon_eps)
+        self.service_eps = []
+
+        self.include_id_paths = False
+        self.get_parameter_values = False
+
+        pars = {}
+        for k, e_v in e.get("example_values", {}).items():
+            pars.update({k: self.argument_definitions[k]})
+            pars[k]["examples"] = e_v
+        for p in self.general_pars:
+            pars.update({p: self.argument_definitions[p]})
+
+
+        self.argument_definitions = pars
+
+
     # -------------------------------------------------------------------------#
 
     def __add_beacon_paths(self):
@@ -109,18 +153,25 @@ class ByconOpenAPI:
         pars = []
         if s == "beaconResultsetsResponse":
             for a, a_d in self.argument_definitions.items():
-                if a_d.get("beacon_query"):
+                if self.mode:
+                    if "VQS" in self.mode:
+                        if a_d.get("vqs_query"):
+                            pars.append(a)
+                    else:
+                        if a_d.get("beacon_query"):
+                            pars.append(a)
+                elif a_d.get("beacon_query"):
                     pars.append(a)
 
         entity_path = f"/{path_part}/{p}"
         self.oapi["paths"].update({
-            entity_path: {"get": self.__path_create_get(r, rqp, r, s, pars)}
+            entity_path: {"get": self.__path_create_get(p, rqp, r, s, pars)}
         })
 
-        if s == "beaconResultsetsResponse":
+        if s == "beaconResultsetsResponse" and self.include_id_paths is True:
             pars = ["id"]
             self.oapi["paths"].update({
-                f"{entity_path}/{{id}}": {"get": self.__path_create_get(r, rqp, r, s, pars)}
+                f"{entity_path}/{{id}}": {"get": self.__path_create_get(p, rqp, r, s, pars)}
             })
 
             # now for the retrieval of the other entities by this id, e.g.
@@ -132,7 +183,7 @@ class ByconOpenAPI:
                 b_p = b_d.get("request_entity_path_id", None)
                 if b_s == "beaconResultsetsResponse" and b_r != r:
                     self.oapi["paths"].update({
-                        f"{entity_path}/{{id}}/{b_p}": {"get": self.__path_create_get(r, rqp, b_r, b_s, pars)}
+                        f"{entity_path}/{{id}}/{b_p}": {"get": self.__path_create_get(p, rqp, b_r, b_s, pars)}
                     })
 
         if s != "beaconInfoResponse":
@@ -142,14 +193,23 @@ class ByconOpenAPI:
 
     # -------------------------------------------------------------------------#
 
-    def __path_create_get(self, source_entity, response_entity_path_alias, target_entity, response_schema, pars):
+    def __path_create_get(self, entity_path_id, response_entity_path_alias, target_entity, response_schema, pars):
+        # using response_entity_path_alias since service responses have special
+        # paths
+        if self.mode:
+            tag = self.mode
+        elif entity_path_id in self.service_eps:
+            tag = "Services"
+        else:
+            tag = "Beacon"
         return {
-            "summary": f"Get {target_entity} entries by {source_entity} id",
-            "description": f"Get {target_entity} entries by {source_entity} id",
+            "summary": f"Get {target_entity} entries",
+            "description": f"Get {target_entity} entries",
+            "tags": [f'{tag}'],
             "parameters": self.__add_parameters(pars, response_entity_path_alias),
             "responses": {
                 "200": {
-                    "description": f"An id response for {source_entity} entry",
+                    "description": f"A response for {target_entity} entries",
                     "content": {
                         "application/json": {
                             "schema": {
@@ -168,7 +228,7 @@ class ByconOpenAPI:
         path_pars = []
         for p in pars:
             if (p_d := self.argument_definitions.get(p)):
-                path_pars.append(self.__format_parameter(camelize(p), p_d, scope))
+                path_pars.append(self.__format_parameter(p, p_d, scope))
         return path_pars
 
 
@@ -176,16 +236,20 @@ class ByconOpenAPI:
 
     def __format_parameter(self, parameter, definition, scope=None):
         p = {
-            "name": parameter,
+            "name": camelize(parameter),
             "in": definition.get("in", "query"),
             "schema": {
                 "type": definition.get("type", "string")
             }
         }
-        if (i := definition.get("items")):
-            p["schema"].update({"items": i})
 
-        if len(e_s := definition.get("examples", [])) > 0:
+        for d_k in ["items", "minItems", "maxItems"]:
+            if (i := definition.get(d_k)):
+                p["schema"].update({d_k: i})
+
+        e_s = definition.get("examples", [])
+        e_s += self.__parameter_get_values(parameter)
+        if len(e_s) > 0:
             p.update({"examples":{}})
             for e in e_s:
                 k = list(e.keys())[0]
@@ -203,10 +267,37 @@ class ByconOpenAPI:
         for a, a_d in self.argument_definitions.items():
             if len(pars) > 0 and a not in pars:
                 continue
-            a_c = camelize(a)
-            c_p = self.__format_parameter(a_c, a_d)
-            self.oapi["components"]["parameters"].update({a_c: c_p})
+            c_p = self.__format_parameter(a, a_d)
+            self.oapi["components"]["parameters"].update({camelize(a): c_p})
 
+
+    # ------------------------------------------------------------------------ #
+
+    def __parameter_get_values(self, par):
+        vals = []
+        if self.get_parameter_values is False:
+            return vals
+        if par in ["reference_name", "mate_name"]:
+            r_l = ChroNames().refseqLabeled()
+            for c in r_l:
+                k = list(c.keys())[0]
+                v = list(c.values())[0]
+                vals.append({v:{"value": k, "summary": f'{v} ({k})'}})
+
+        elif par in ["variant_type"]:
+            for v_t, v_d in BYC.get("variant_type_definitions").items():
+                if not v_d.get("beacon_query", True):
+                    continue
+                vals.append({v_t:{"value": v_t, "summary": f'{v_t} ({v_d.get("variant_state", {}).get("label", "")})'}})
+
+        elif par == "filters":
+            BYC.update({"response_entity_id": "filteringTerm", "TEST_MODE": True})
+            BYC_PARS.update({"test_mode_count": 10})
+            f_t_s = ByconFilteringTerms().filteringTermsList()
+            for f in f_t_s:
+                vals.append({f["id"]:{"value": [f["id"]], "summary": f'{f["id"]}: {f['label']}'}})
+
+        return vals
 
 # def __path_create_post(source_entity, response_entity_path_alias, target_entity, this_server, response_schema, par_def):
 #     return  # {